Boatload of changes/fixes
- Added checkpoint at block 50,000
- Sealed Network transition by permanently setting Block version to 2
and rejecting all Wallets version 0.4.x
- Reduced Reward to 25 KHC after block 250,000 and reduced flat fee to
0.2%.  Minimum flat fee increased to 10 kryptoha-toshis.
- Improved the calculation of the flat fee.
- Allow already processed TXs with timestamp older than 2 weeks or, with
incorrectly calculated flat fee to enter the Mempool when the Wallet is
re-indexing, importing or downloading blocks.

+++ src/qt/coincontroldialog.cpp
@@ -540,8 +540,8 @@ void CoinControlDialog::updateLabels(WalletModel *model, QDialog* dialog)
         // Fee
         int64_t nFee = nTransactionFee * (1 + (int64_t)nBytes / 1000);
         // Flat Fee of 0.5%
-        int64_t nFlatFee = (int64_t)(nPayAmount * FLAT_FEE_PER_TRANSACTION) > MIN_FLAT_TRANSACTION_FEE ?
-                           (int64_t)(nPayAmount * FLAT_FEE_PER_TRANSACTION) : MIN_FLAT_TRANSACTION_FEE;
+        int64_t nFlatFee = (int64_t)((double)nPayAmount * FLAT_FEE_PER_TRANSACTION) > MIN_FLAT_TRANSACTION_FEE ?
+                           (int64_t)((double)nPayAmount * FLAT_FEE_PER_TRANSACTION) : MIN_FLAT_TRANSACTION_FEE;
 
         // Min Fee
         int64_t nMinFee = GetMinFee(txDummy, nBytes, AllowFree(dPriority), GMF_SEND);
